+++ Changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 5.1.0 - 2021-09-23
+### Changed
+- The license check now ignores all files in `.gitignore` and the folder
+  `.git`. Before it just ignored `node_modules`, `dist`, and `.git`.
+### Fixed
+- With `npm run lint` the license check failed to run for shared on Windows.
+
 ## 5.0.1 - 2021-09-17
 ### Fixed
 - Format of library version from nrf-device-lib-js.